+++ tests/utils/mixins.py
@@ -1,4 +1,3 @@
-import pulsar
 from pulsar.apps.test import unittest
 
 
@@ -6,7 +5,8 @@ class TestMixins(unittest.TestCase):
     
     def testLocal(self):
         from pulsar.utils.structures import AttributeDictionary
-        elem = pulsar.LocalMixin()
+        from pulsar.utils.log import LocalMixin
+        elem = LocalMixin()
         el = elem.local
         self.assertTrue(isinstance(el, AttributeDictionary))
         self.assertEqual(id(elem.local), id(el))